Extract FormField into a standalone reusable component.
